Add files via upload
Project Background 

TikTok’s data team is in the earliest stages of the claims classification project. The following tasks are needed before the team can begin the data analysis process:

Build a dataframe for the TikTok dataset

Examine data type of each column

Gather descriptive statistics

Your assignment
You will build a dataframe for the claims classification data. After the dataframe is complete, you will organize the claims data for the process of exploratory data analysis, and update the team on your progress and insights.